# src/main/resources/api-swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"Api Documentation",
    "version":"1.0",
    "title":"Api Documentation",
    "termsOfService":"urn:tos",
    "contact":{

    },
    "license":{
      "name":"Apache 2.0",
      "url":"http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host":"localhost:8080",
  "basePath":"/",
  "tags":[
    {
      "name":"payment-analytic-controller",
      "description":"Payment Analytic Controller"
    }
  ],
  "paths":{
    "/analytic":{
      "get":{
        "tags":[
          "payment-analytic-controller"
        ],
        "summary":"getAllAnalytic",
        "operationId":"getAllAnalyticUsingGET",
        "produces":[
          "application/json"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"array",
              "items":{
                "$ref":"#/definitions/UserPaymentAnalytic"
              }
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        },
        "deprecated":false
      }
    },
    "/analytic/{userId}":{
      "get":{
        "tags":[
          "payment-analytic-controller"
        ],
        "summary":"getUserAnalytic",
        "operationId":"getUserAnalyticUsingGET",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"userId",
            "in":"path",
            "description":"userId",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/UserPaymentAnalytic"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        },
        "deprecated":false
      }
    },
    "/analytic/{userId}/stats":{
      "get":{
        "tags":[
          "payment-analytic-controller"
        ],
        "summary":"getPaymentStatsByUser",
        "operationId":"getPaymentStatsByUserUsingGET",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"userId",
            "in":"path",
            "description":"userId",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/UserPaymentStats"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        },
        "deprecated":false
      }
    },
    "/analytic/{userId}/templates":{
      "get":{
        "tags":[
          "payment-analytic-controller"
        ],
        "summary":"getUserTemplates",
        "operationId":"getUserTemplatesUsingGET",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"userId",
            "in":"path",
            "description":"userId",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"array",
              "items":{
                "$ref":"#/definitions/UserTemplate"
              }
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        },
        "deprecated":false
      }
    }
  },
  "definitions":{
    "PaymentCategoryInfo":{
      "type":"object",
      "properties":{
        "max":{
          "type":"number"
        },
        "min":{
          "type":"number"
        },
        "sum":{
          "type":"number"
        }
      },
      "title":"PaymentCategoryInfo"
    },
    "UserPaymentAnalytic":{
      "type":"object",
      "properties":{
        "analyticInfo":{
          "type":"object",
          "additionalProperties":{
            "$ref":"#/definitions/PaymentCategoryInfo"
          }
        },
        "totalSum":{
          "type":"number"
        },
        "userId":{
          "type":"string"
        }
      },
      "title":"UserPaymentAnalytic"
    },
    "UserPaymentStats":{
      "type":"object",
      "properties":{
        "maxAmountCategoryId":{
          "type":"integer",
          "format":"int32"
        },
        "minAmountCategoryId":{
          "type":"integer",
          "format":"int32"
        },
        "oftenCategoryId":{
          "type":"integer",
          "format":"int32"
        },
        "rareCategoryId":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"UserPaymentStats"
    },
    "UserTemplate":{
      "type":"object",
      "properties":{
        "amount":{
          "type":"number"
        },
        "categoryId":{
          "type":"integer",
          "format":"int32"
        },
        "recipientId":{
          "type":"string"
        }
      },
      "title":"UserTemplate"
    }
  }
}